refactor: integration service - mining relate cases
