+++ test/src/specs/mining/proposal.rs
@@ -1,6 +1,5 @@
 use crate::{Net, Spec};
 use ckb_types::prelude::*;
-use log::info;
 
 pub struct AvoidDuplicatedProposalsWithUncles;
 
@@ -10,18 +9,15 @@ impl Spec for AvoidDuplicatedProposalsWithUncles {
     // Case: This is not a validation rule, but just an improvement for miner
     //       filling proposals: Don't re-propose the transactions which
     //       has already been proposed within the uncles.
+    //    1. Submit `tx` into mempool, and `uncle` which proposed `tx` as an candidate uncle
+    //    2. Get block template, expect empty proposals cause we already proposed `tx` within `uncle`
+
     fn run(&self, net: &mut Net) {
         let node = &net.nodes[0];
-        let window = node.consensus().tx_proposal_window();
-        node.generate_blocks(window.farthest() as usize + 2);
+        node.generate_blocks_until_contains_valid_cellbase();
 
-        info!(
-            "(1) Submit `tx` into mempool, and `uncle` which proposed `tx` as an candidate uncle"
-        );
-        let tx = {
-            node.generate_block();
-            node.new_transaction_spend_tip_cellbase()
-        };
+        let tx = node.new_transaction_spend_tip_cellbase();
+
         let uncle = {
             let block = node.new_block(None, None, None);
             let uncle = block
@@ -35,9 +31,6 @@ impl Spec for AvoidDuplicatedProposalsWithUncles {
         node.submit_block(&uncle);
         node.submit_transaction(&tx);
 
-        info!(
-            "(2) Get block template, expect empty proposals cause we already proposed `tx` within `uncle`"
-        );
         let block = node.new_block(None, None, None);
         assert_eq!(
             vec![uncle.hash()],
