+++ test/src/specs/mining/tx_select.rs
@@ -1,10 +1,6 @@
-use crate::{Net, Node, Spec, DEFAULT_TX_PROPOSAL_WINDOW};
-use ckb_types::{
-    bytes::Bytes,
-    core::{Capacity, TransactionView},
-    prelude::*,
-};
-use log::info;
+use crate::util::tx::new_transaction_with_fee_and_size;
+use crate::{Net, Spec, DEFAULT_TX_PROPOSAL_WINDOW};
+use ckb_types::core::Capacity;
 
 pub struct TemplateTxSelect;
 
@@ -12,21 +8,17 @@ impl Spec for TemplateTxSelect {
     crate::name!("template_tx_select");
 
     fn run(&self, net: &mut Net) {
-        self.select_higher_tx_fee(net);
-    }
-}
-
-impl TemplateTxSelect {
-    fn select_higher_tx_fee(&self, net: &mut Net) {
         let node = &net.nodes[0];
         // prepare blocks
         node.generate_blocks((DEFAULT_TX_PROPOSAL_WINDOW.1 + 6) as usize);
-        let mut txs_hash = Vec::new();
-        let block = node.get_tip_block();
-        let number = block.header().number();
+        let number = node.get_tip_block_number();
+        let blank_block_size = node
+            .get_tip_block()
+            .data()
+            .serialized_size_without_uncle_proposals();
 
-        info!("Generate txs");
-        // send 5 txs which tx fee rate is same
+        // send 3 txs which tx fee rate is same
+        let mut txs_hash = Vec::new();
         [501, 501, 501, 501, 300]
             .iter()
             .enumerate()
@@ -44,40 +36,14 @@ impl TemplateTxSelect {
             });
 
         // skip proposal window
-        node.generate_block();
-        node.generate_block();
+        node.generate_blocks(DEFAULT_TX_PROPOSAL_WINDOW.0 as usize);
 
-        let new_block = node.new_block(Some(1000), None, None);
+        let new_block = node.new_block(Some(blank_block_size as u64 + 900), None, None);
         // should choose two txs: 501, 300
-        assert_eq!(new_block.transactions().len(), 2);
+        assert_eq!(
+            new_block.transactions().len(),
+            3,
+            "New block should contain txs: 501, 300"
+        );
     }
-}
-
-fn new_transaction_with_fee_and_size(
-    node: &Node,
-    parent_tx: &TransactionView,
-    fee: Capacity,
-    tx_size: usize,
-) -> TransactionView {
-    let input_capacity: Capacity = parent_tx
-        .outputs()
-        .get(0)
-        .expect("parent output")
-        .capacity()
-        .unpack();
-    let capacity = input_capacity.safe_sub(fee).unwrap();
-    let tx = node.new_transaction_with_since_capacity(parent_tx.hash(), 0, capacity);
-    let original_tx_size = tx.data().serialized_size_in_block();
-    let tx = tx
-        .as_advanced_builder()
-        .set_outputs_data(vec![
-            Bytes::from(vec![0u8; tx_size - original_tx_size]).pack()
-        ])
-        .build();
-    assert_eq!(
-        tx.data().serialized_size_in_block(),
-        tx_size,
-        "tx size incorrect"
-    );
-    tx
 }
